@@ -11,17 +11,21 @@
   },
   "private": true,
   "dependencies": {
+    "@angular/animations": "^16.2.11",
     "@angular/cdk": "^16.2.11",
     "@angular/common": "^16.1.0",
     "@angular/compiler": "^16.1.0",
-    "@angular/core": "^16.1.0",
+    "@angular/core": "^16.2.12",
     "@angular/forms": "^16.1.0",
     "@angular/google-maps": "^16.0.0",
+    "@angular/material": "^16.2.11",
     "@angular/platform-browser": "^16.1.0",
     "@angular/platform-browser-dynamic": "^16.1.0",
     "@angular/router": "^16.1.0",
+    "@auth0/angular-jwt": "^5.2.0",
     "@coreui/angular": "^4.5.27",
     "@coreui/icons-angular": "^4.5.27",
+    "@costlydeveloper/ngx-awesome-popup": "^3.2.0",
     "@ng-bootstrap/ng-bootstrap": "^15.1.2",
     "@popperjs/core": "^2.11.6",
     "bootstrap": "^5.3.2",
